+++ front/main/app/components/post-reply.js
@@ -1,13 +1,14 @@
 import Ember from 'ember';
 import DiscussionParsedContentMixin from '../mixins/discussion-parsed-content';
+import DiscussionMoreOptionsMixin from '../mixins/discussion-more-options';
 
 export default Ember.Component.extend(
 	DiscussionParsedContentMixin,
+	DiscussionMoreOptionsMixin,
 	{
 		classNames: ['post-reply'],
 		classNameBindings: ['isNew', 'isDeleted', 'isParentDeleted'],
 
-		isDeleted: Ember.computed.alias('post.isDeleted'),
-		post: null
+		isDeleted: Ember.computed.alias('post.isDeleted')
 	}
 );