+++ front/main/app/components/side-nav.js
@@ -6,92 +6,68 @@ export default Ember.Component.extend({
 	classNames: ['side-nav'],
 	classNameBindings: ['shouldBeVisible:slide-into-view:collapsed'],
 
-	isInSearchMode: false,
-	searchQuery: '',
-	searchPlaceholderLabel: Ember.computed(() => {
-		return i18n.t('app.search-label');
-	}),
-
 	shouldBeVisibleObserver: Ember.observer('shouldBeVisible', function () {
-		const trackLabel = this.get('shouldBeVisible') ? 'open' : 'close';
-
 		track({
 			action: trackActions.click,
-			category: 'menu',
-			label: trackLabel
+			category: 'wiki-nav',
+			label: this.get('shouldBeVisible') ? 'open' : 'close'
 		});
 	}),
 
-	/**
-	 * Every time we exit search mode, regardless of if it was through the Cancel
-	 * link or through clicking a search result, we want to clear out the query
-	 * so that the search bar will clear.
-	 */
-	isInSearchModeObserver: Ember.observer('isInSearchMode', function () {
-		if (!this.get('isInSearchMode')) {
-			this.send('clearSearch');
-		}
-	}).on('didInsertElement'),
+	globalNavContent: 'side-nav-global-navigation-root',
+	isFandomVisible: Ember.computed(() => Mercury.wiki.language.content === 'en'),
+	wikiaHomepage: Ember.getWithDefault(Mercury, 'wiki.homepage', 'http://www.wikia.com'),
+	homeOfFandomLabel: Ember.get(Mercury, 'wiki.navigation2016.fandomLabel'),
 
 	actions: {
 		/**
 		 * @returns {void}
 		 */
-		clearSearch() {
-			this.set('searchQuery', '');
+		wordmarkClick() {
+			track({
+				action: trackActions.click,
+				category: 'wiki-nav',
+				label: 'wordmark',
+			});
+			this.send('collapse');
+		},
+
+		/**
+		 * @returns {void}
+		 */
+		homeOfFandomClick() {
+			track({
+				action: trackActions.click,
+				category: 'wiki-nav',
+				label: 'home-of-fandom',
+			});
+			this.send('collapse');
 		},
 
 		/**
 		 * @returns {void}
 		 */
 		collapse() {
+			this.set('globalNavContent', 'side-nav-global-navigation-root');
 			this.sendAction('toggleVisibility', false);
-			this.send('searchCancel');
-		},
-
-		/**
-		 * @returns {void}
-		 */
-		expand() {
-			this.sendAction('toggleVisibility', true);
-		},
-
-		/**
-		 * @returns {void}
-		 */
-		searchCancel() {
-			this.set('isInSearchMode', false);
-			this.send('clearSearch');
-		},
-
-		/**
-		 * @returns {void}
-		 */
-		searchFocus() {
-			this.set('isInSearchMode', true);
-			// Track when search is opened
-			track({
-				action: trackActions.click,
-				category: 'search',
-			});
 		},
 
 		/**
 		 * @returns {void}
 		 */
 		loadRandomArticle() {
+			this.set('globalNavContent', 'side-nav-global-navigation-root');
 			this.sendAction('loadRandomArticle');
 		},
 
-		/**
-		 * Handler for enter in search box
-		 *
-		 * @param {string} [value=''] - input value
-		 * @returns {void}
-		 */
-		enter(value) {
-			// Use Wikia Search
-			window.location.assign(`${Mercury.wiki.articlePath}Special:Search?search=${value}&fulltext=Search`);
-		},
+		replaceNavigationContent(navName) {
+			if (navName === 'explore') {
+				this.set('globalNavContent', 'side-nav-explore-wikia-navigation');
+			} else if (navName === 'local') {
+				this.set('globalNavContent', 'side-nav-local-navigation-root');
+			} else if (navName === 'root') {
+				this.set('globalNavContent', 'side-nav-global-navigation-root');
+			}
+		}
 	},
 });