+++ front/main/app/routes/discussion/base.js
@@ -1,5 +1,6 @@
 import Ember from 'ember';
 import {updateTrackedUrl, trackPageView} from 'common/utils/track';
+import UniversalAnalytics from 'common/modules/Trackers/UniversalAnalytics';
 
 export default Ember.Route.extend({
 	postDeleteFullScreenOverlay: false,
@@ -13,18 +14,12 @@ export default Ember.Route.extend({
 		},
 
 		/**
-		 * @returns {void}
-		 */
-		goToAllDiscussions() {
-			this.transitionTo('discussion.index');
-		},
-
-		/**
 		 * @returns {boolean}
 		 */
 		didTransition() {
 			this.controllerFor('application').set('noMargins', true);
 
+			UniversalAnalytics.setDimension(19, null);
 			updateTrackedUrl(window.location.href);
 			trackPageView(null);