+++ gulp/paths.js
@@ -22,7 +22,8 @@ module.exports = {
 			src: 'front/auth/app/styles',
 			dest: outputFront + '/auth/assets',
 			partials: '_*.scss',
-			compile: '*.scss'
+			compile: '*.scss',
+			watch: 'front/auth/app/styles/**/*.scss'
 		},
 		vendor: {
 			src: 'front/auth/bower_components/**/*',
@@ -89,8 +90,7 @@ module.exports = {
 			main: {
 				src: outputFront + '/main/index.html',
 				dest: outputServer + '/app/views/_layouts',
-				outputFilename: 'ember-main.hbs',
-				watch: outputFront + '/**/*'
+				outputFilename: 'ember-main.hbs'
 			}
 		},
 	},