+++ homepage/front/scripts/main/main.js
@@ -1,10 +1,13 @@
-import {loadGlobalData, getLoginUrl, getJaCommunityUrl, getJaUniversityUrl, getStartWikiaUrl} from './globals';
+import * as globals from './globals';
+import {loadSearch} from './search';
 
 /**
+ * Perform search
  * @returns {void}
  */
 function search(isTopNav = true) {
-	let searchText;
+	let searchText = encodeURI($('#searchWikiaText').val()),
+		searchUrl;
 
 	if (isTopNav) {
 		searchText = encodeURI($('#searchWikiaText').val());
@@ -23,15 +26,84 @@ function search(isTopNav = true) {
 	}
 
 	if (searchText) {
-		window.location.href = `http://ja.wikia.com/Special:Search?search=${searchText}&fulltext=Search&resultsLang=ja`;
+		if (window.optimizely.variationMap[globals.getOptimizelyId()] === 1) {
+			// Use Google search
+			if (window.location.pathname === '/') {
+				searchUrl = `search?q=${searchText}`;
+			} else {
+				searchUrl = `/search?q=${searchText}`;
+			}
+		} else {
+			// Use Oasis search
+			searchUrl = `http://ja.wikia.com/Special:Search?search=${searchText}&fulltext=Search&resultsLang=ja`;
+		}
+
+		window.location.href = searchUrl;
+	}
+}
+
+/**
+ * Hides the loading indicator
+ * @returns {void}
+ */
+function hideLoadingIndicator() {
+	$('#loading').addClass('loading-done');
+}
+
+/**
+ * Calculates and sets compensation for the right slider arrow to account for variable spacing between slides.
+ * See INT-319 and INT-329. 250 is the fixed pixel width of each slide
+ * @returns {void}
+ */
+function calculateCarouselCompensation() {
+	const delta = $('#carousel-1 .slick-slide').width() - 250,
+		deltaMedium = $('#carousel-1-medium .slick-slide').width() - 250;
+
+	for (let i = 1; i <= 5; i++) {
+		$(`#carousel-${i} .slick-prev`).addClass('slick-prev-category');
+		$(`#carousel-${i}-medium .slick-prev`).addClass('slick-prev-category');
+
+		if (delta > 0) {
+			$(`#carousel-${i} .slick-next`).css('right', delta + 10);
+		} else {
+			$(`#carousel-${i} .slick-next`).css('right', '');
+		}
+
+		if (deltaMedium > 0) {
+			$(`#carousel-${i}-medium .slick-next`).css('right', deltaMedium + 10);
+		} else {
+			$(`#carousel-${i}-medium .slick-next`).css('right', 20);
+		}
+	}
+
+	if (delta > 0) {
+		// Compensation for slide not being fully centered due to responsive slider spacing
+		$('.featured').css('padding-left', 54 + delta);
+	} else {
+		$('.featured').css('padding-left', '');
+	}
+
+	if (deltaMedium > 0) {
+		// Compensation for slide not being fully centered due to responsive slider spacing
+		$('.featured-mobile-medium').css('padding-left', 54 + deltaMedium);
+	} else {
+		$('.featured-mobile-medium').css('padding-left', '');
 	}
 }
 
 $(() => {
-	const headings = $('.grid-heading');
+	// Hide loading indicator after load complete
+	$(window).load(() => {
+		hideLoadingIndicator();
+	});
+
+	// Or after 6 seconds
+	setTimeout(() => {
+		hideLoadingIndicator();
+	}, 6000);
 
 	$('.hero-carousel').slick({
-		arrows: false,
+		arrows: true,
 		dots: true,
 		autoplay: true,
 		autoplaySpeed: 3000,
@@ -47,64 +119,36 @@ $(() => {
 	});
 
 	$('.hero-carousel-mobile').slick({
-		arrows: false,
+		arrows: true,
 		dots: true,
 		autoplay: true,
 		autoplaySpeed: 3000,
 		slidesToShow: 1,
 	});
 
-	// Move previous/next arrow elements inside hero-carousel.
-	// This must be done after initializing slick, otherwise the buttons will
-	// be treated as slides
-	$('#hero-prev').detach().appendTo('#hero-carousel');
-	$('#hero-next').detach().appendTo('#hero-carousel');
-	$('#hero-prev-mobile').detach().appendTo('#hero-carousel-mobile');
-	$('#hero-next-mobile').detach().appendTo('#hero-carousel-mobile');
-
 	$('.featured-carousel').slick({
-		arrows: false,
+		arrows: true,
 		dots: false,
 		slidesToShow: 4,
 		slidesToScroll: 4,
-		responsive: [
-			{
-				breakpoint: 1140,
-				settings: {
-					slidesToShow: 3,
-					slidesToScroll: 3,
-				}
-			},
-			{
-				breakpoint: 865,
-				settings: {
-					slidesToShow: 2,
-					slidesToScroll: 2
-				}
-			},
-			{
-				breakpoint: 615,
-				settings: {
-					slidesToShow: 1,
-					slidesToScroll: 1
-				}
-			}
-		]
 	});
 
-	// TODO: The number of carousels should be encoded in a json file
-	for (let i = 1; i <= 5; i++) {
-		$(`#carousel-${i}-prev`).detach().appendTo(`#carousel-${i}`);
-		$(`#carousel-${i}-next`).detach().appendTo(`#carousel-${i}`);
-	}
+	$('.featured-carousel-medium').slick({
+		arrows: true,
+		dots: false,
+		slidesToShow: 3,
+		slidesToScroll: 3,
+	});
 
-	// Dynamically adjust text size to show community title without text break.
-	// bigText adjusts the size programatically and strips off css padding, so it is
-	// necessary to add it in explicitly afterwards
-	headings.bigText({maximumFontSize: 20, verticalAlign: 'top'});
-	headings.css({padding: '.1rem'});
+	$(window).resize(() => {
+		calculateCarouselCompensation();
+	});
 
-	loadGlobalData();
+	calculateCarouselCompensation();
+
+	globals.loadGlobalData().then((data) => {
+		loadSearch(data.mobileBreakpoint);
+	});
 });
 
 $('#beginnersGuide').click((event) => {
@@ -123,11 +167,11 @@ $('.search-wikia').click((event) => {
 });
 
 $('.jw-community-link').click(() => {
-	window.location.href = getJaCommunityUrl();
+	window.location.href = globals.getJaCommunityUrl();
 });
 
 $('.jw-university-link').click(() => {
-	window.location.href = getJaUniversityUrl();
+	window.location.href = globals.getJaUniversityUrl();
 });
 
 $('.wiw-search-wikia-form').submit((event) => {
@@ -141,34 +185,15 @@ $('.wiw-search-wikia-button').click((event) => {
 });
 
 $('.wiw-start-wikia').click(() => {
-	window.location.href = getStartWikiaUrl();
-});
-
-$('.hero-prev').click(function () {
-	const id = $(this).attr('id'),
-		carousel = $(`#${id}`).parent().attr('id');
-
-	$(`#${carousel}`).slick('slickPrev');
-});
-
-$('.hero-next').click(function () {
-	const id = $(this).attr('id'),
-		carousel = $(`#${id}`).parent().attr('id');
-
-	$(`#${carousel}`).slick('slickNext');
+	window.location.href = globals.getStartWikiaUrl();
 });
 
 $('#loginIcon').click((event) => {
-	if ($(document).width() < 710) {
+	if ($(document).width() < globals.getMobileBreakpoint()) {
 		$('#userInfoToggle').toggle();
 	} else {
-		window.location.href = getLoginUrl();
+		window.location.href = globals.getLoginUrl();
 	}
 
-	event.preventDefault();
-});
-
-$('#whatIsWikia').click((event) => {
-	window.location.href = '/beginners';
 	event.preventDefault();
 });